Hacky bug fixes after upddating the fxn to get spikes
I had changed the function that extracts spikes to make it work
with single electrodes, tetrodes, whatever. This broke the
code for getting the spikes from a cluster. I made a quick fix
to this function to get it to work but it will only work for clusters
that were recorded using tetrodes now. The underlying problem that needs
to be fixed is that we need to specify the type of electrode that was
used in the cluster database, so that later we can use it to get the
correct spikes file.